initial contruction on naming refactorization
